LibWeb: Split BackgroundRepeatStyleValue out of StyleValue.{h,cpp}
