+++ Userland/Libraries/LibWeb/CMakeLists.txt
@@ -65,6 +65,7 @@ set(SOURCES
     CSS/StyleSheetList.cpp
     CSS/StyleValue.cpp
     CSS/StyleValues/AngleStyleValue.cpp
+    CSS/StyleValues/BackgroundRepeatStyleValue.cpp
     CSS/StyleValues/BackgroundStyleValue.cpp
     CSS/Supports.cpp
     CSS/SyntaxHighlighter/SyntaxHighlighter.cpp
