+++ Userland/Libraries/LibWeb/Layout/Node.cpp
@@ -5,6 +5,7 @@
  */
 
 #include <AK/Demangle.h>
+#include <LibWeb/CSS/StyleValues/BackgroundRepeatStyleValue.h>
 #include <LibWeb/DOM/Document.h>
 #include <LibWeb/Dump.h>
 #include <LibWeb/HTML/BrowsingContext.h>
